requirements.txt
# QualityCore Inventory Backend - Dependencies
# Using version ranges to allow security patches while maintaining compatibility

# Core Framework
Django>=5.2.2,<6.0
djangorestframework>=3.16.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0

# Production Server
gunicorn>=22.0.0,<23.0

# Database
psycopg2-binary>=2.9.10,<3.0

# CORS Headers
django-cors-headers>=4.7.0,<5.0

# Environment Variables
django-environ>=0.12.0,<1.0

# Image Storage (Cloudinary)
cloudinary>=1.41.0,<2.0
django-cloudinary-storage>=0.3.0,<1.0
Pillow>=11.2.1,<12.0

# Async Tasks (Celery + Redis)
celery>=5.6.0,<6.0
redis>=7.1.0,<8.0

# PDF Generation
reportlab>=4.4.2,<5.0

# HTTP Requests & Utilities
requests>=2.32.5,<3.0
python-dateutil>=2.9.0,<3.0

# API Documentation (Swagger/OpenAPI)
drf-spectacular>=0.29.0,<1.0

# Error Monitoring (Sentry)
sentry-sdk>=2.20.0,<3.0

# Structured Logging (JSON)
python-json-logger>=3.2.0,<4.0

# Celery Dependencies
amqp>=5.3.1,<6.0
asgiref>=3.8.1,<4.0
billiard>=4.2.4,<5.0
certifi>=2025.11.12
charset-normalizer>=3.4.2,<4.0
click>=8.3.1,<9.0
click-didyoumean>=0.3.1,<1.0
click-plugins>=1.1.1,<2.0
click-repl>=0.3.0,<1.0
colorama>=0.4.6,<1.0
exceptiongroup>=1.3.1,<2.0
idna>=3.11,<4.0
kombu>=5.6.1,<6.0
packaging>=25.0,<26.0
prompt_toolkit>=3.0.52,<4.0
six>=1.17.0,<2.0
sqlparse>=0.5.3,<1.0
tzdata>=2025.2
tzlocal>=5.3.1,<6.0
urllib3>=2.6.1,<3.0
vine>=5.1.0,<6.0
wcwidth>=0.2.14,<1.0
